pyproject.toml
[tool.poetry]
name = "kintai-bot-python"
version = "0.1.0"
description = ""
authors = ["t-makino <t-makino@atware.co.jp>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"
gspread = "^3.6.0"
oauth2client = "^4.1.3"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
